backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-jwt-extended
marshmallow
flask-marshmallow
pydantic
pymysql
cryptography
python-dotenv
pytest
